# backend/core/docs/swagger.yaml
basePath: /
definitions:
...
  handler.AnalyticsSummaryResponseDTO:
    properties:
      end_date:
        type: string
      global_revenue:
        type: number
      global_tickets:
        type: integer
      start_date:
        type: string
      stats_by_cinema:
        items:
          $ref: '#/definitions/handler.DailyCinemaStatsResponseDTO'
        type: array
    type: object
...
  handler.DailyCinemaStatsResponseDTO:
    properties:
      cinema_name:
        type: string
      date:
        type: string
      occupancy_rate:
        type: number
      tickets_sold:
        type: integer
      total_revenue:
        type: number
    type: object
...
  handler.GenreStatsResponseDTO:
    properties:
      genre_name:
        type: string
      total_revenue:
        type: number
    type: object
...
  handler.MovieStatsDTO:
    properties:
      movie_id:
        type: integer
      movie_title:
        type: string
      tickets_sold:
        type: integer
      total_revenue:
        type: number
    type: object
...
  httputil.MessageResponse:
    properties:
      message:
        type: string
    type: object
  internal_auth_handler.UpdateRoleDTO:
    properties:
      role:
        enum:
        - USER
        - MANAGER
        - ADMIN
        type: string
    required:
    - role
    type: object
...
host: localhost:8003
info:
  contact:
    email: suporte@screek.com
    name: Suporte screeK
    url: http://www.screek.com/support
  description: API para a plataforma de cinema screeK.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: screeK API
  version: "1.0"
paths:
  /admin/analytics/genres:
    get:
      description: Analisa a distribuição de vendas por gênero de filme
      parameters:
      - description: Data inicial
        in: query
        name: start
        type: string
      - description: Data final
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.GenreStatsResponseDTO'
            type: array
      security:
      - BearerAuth: []
      summary: Preferência por gênero
      tags:
      - Analytics
  /admin/analytics/movies:
    get:
      description: Retorna ranking de filmes por bilheteria e ingressos vendidos
      parameters:
      - description: Data inicial
        in: query
        name: start
        type: string
      - description: Data final
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.MovieStatsDTO'
            type: array
      security:
      - BearerAuth: []
      summary: Desempenho por filme
      tags:
      - Analytics
  /admin/analytics/revenue:
    get:
      description: Retorna métricas financeiras (total, média, tendências)
      parameters:
      - description: Data inicial (YYYY-MM-DD)
        in: query
        name: start
        type: string
      - description: Data final (YYYY-MM-DD)
        in: query
        name: end
        type: string
      - description: Periodicidade (daily, monthly)
        in: query
        name: period
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AnalyticsSummaryResponseDTO'
      security:
      - BearerAuth: []
      summary: Estatísticas de receita
      tags:
      - Analytics
...
  /admin/users/{id}/role:
    patch:
      consumes:
      - application/json
      description: Altera o nível de acesso (Role) de um usuário (Apenas Admin)
      parameters:
      - description: ID do Usuário (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Novo Papel
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth_handler.UpdateRoleDTO'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputil.MessageResponse'
      summary: Atualizar papel do usuário
      tags:
      - Auth (Admin)
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"